feat(cli): update Metro stack frame filtering
