+++ packages/@expo/cli/src/start/server/metro/__tests__/metroErrorInterface.test.ts
@@ -1,7 +1,7 @@
+import { likelyContainsCodeFrame } from '../formatStack';
 import {
   attachImportStackToRootMessage,
   nearestImportStack,
-  likelyContainsCodeFrame,
   dropStackIfContainsCodeFrame,
 } from '../metroErrorInterface';
