+++ packages/@expo/metro-config/src/customizeFrame.ts
@@ -10,35 +10,44 @@ type CustomizeFrameFunc = SymbolicatorConfigT['customizeFrame'];
 // Import only the types here, the values will be imported from the project, at runtime.
 export const INTERNAL_CALLSITES_REGEX = new RegExp(
   [
-    '/Libraries/Renderer/implementations/.+\\.js$',
+    // React Native internal callsites:
+    // https://github.com/react/react-native/blob/f2a250ad895ffd4478f0668f56867b78be21ea2f/packages/metro-config/src/index.flow.js#L20-L48
     '/Libraries/BatchedBridge/MessageQueue\\.js$',
-    '/Libraries/YellowBox/.+\\.js$',
+    '/Libraries/Core/.+\\.js$',
     '/Libraries/LogBox/.+\\.js$',
-    '/Libraries/Core/Timers/.+\\.js$',
-    'node_modules/react-devtools-core/.+\\.js$',
-    'node_modules/react-refresh/.+\\.js$',
-    'node_modules/scheduler/.+\\.js$',
+    '/Libraries/Network/.+\\.js$',
+    '/Libraries/Pressability/.+\\.js$',
+    '/Libraries/Renderer/implementations/.+\\.js$',
+    '/Libraries/Utilities/.+\\.js$',
+    '/Libraries/vendor/.+\\.js$',
+    '/Libraries/WebSocket/.+\\.js$',
+    '/src/private/renderer/errorhandling/.+\\.js$',
+    '/metro-runtime/.+\\.js$',
+    '/node_modules/@babel/runtime/.+\\.js$',
+    '/node_modules/@react-native/js-polyfills/.+\\.js$',
+    '/node_modules/invariant/.+\\.js$',
+    '/node_modules/react-devtools-core/.+\\.js$',
+    '/node_modules/react-native/index.js$',
+    '/node_modules/react-refresh/.+\\.js$',
+    '/node_modules/scheduler/.+\\.js$',
+
+    `\\[native code\\]`, // relaxed compared to '^\\[native code\\]$'
+
+    '/Libraries/YellowBox/.+\\.js$',
     // Metro replaces `require()` with a different method,
     // we want to omit this method from the stack trace.
     // This is akin to most React tooling.
     '/metro/.*/polyfills/require.js$',
     // Hide frames related to a fast refresh.
     '/metro/.*/lib/bundle-modules/.+\\.js$',
-    'node_modules/react-native/Libraries/Utilities/HMRClient.js$',
     'node_modules/eventemitter3/index.js',
     'node_modules/event-target-shim/dist/.+\\.js$',
-    // Improve errors thrown by invariant (ex: `Invariant Violation: "main" has not been registered`).
-    'node_modules/invariant/.+\\.js$',
     // Remove babel runtime additions
     'node_modules/regenerator-runtime/.+\\.js$',
     // Remove react native setImmediate ponyfill
     'node_modules/promise/setimmediate/.+\\.js$',
-    // Babel helpers that implement language features
-    'node_modules/@babel/runtime/.+\\.js$',
     // Hide Hermes internal bytecode
     '/(?:InternalBytecode/)?InternalBytecode\\.js$',
-    // Block native code invocations
-    `\\[native code\\]`,
     // Hide react-dom (web)
     'node_modules/react-dom/.+\\.js$',
     // Hide node.js evaluation code
@@ -47,8 +56,6 @@ export const INTERNAL_CALLSITES_REGEX = new RegExp(
     '@expo/metro-runtime/.+\\.ts',
     '@expo/server/.+\\.ts',
     'expo-server/.+\\.ts',
-    // Block upstream metro-runtime
-    '/metro-runtime/.+\\.js$',
     // Expo's metro-runtime require patch:
     '@expo/metro-config/require/.+',
 
@@ -64,9 +71,6 @@ export const INTERNAL_CALLSITES_REGEX = new RegExp(
     '/react-native-web/dist/.+\\.js$',
     // React Server Components adapter (note we should probably use an Expo-Metro-specific version in the future).
     'node_modules/react-server-dom-webpack/.+\\.js$',
-
-    // Block all node modules.
-    'node_modules/.+/',
   ].join('|')
 );
